added indexer and fixed gitignore

+++ cleanup.py
@@ -2,4 +2,6 @@ import os
 def clean():
     x = os.listdir("dbg")
     for i in x:
-        os.remove(f"dbg/{i}")
+        print(i)
+        if i != "gitinclude":
+            os.remove(f"dbg/{i}")
